Add CONDOR.a.f
CONDOR.a.f is a Reduced/Revised version of CONDON, to be linked and compiled with a fortran95 compiler, to be linked with a diagonalization routine for Hermitian matrices, e.g. dsdzhd.f! CONDOR works for "imaginary" Wybourne values, too, but many options are missing. Step-operators are used throughout in new subroutines.